fix: update model standings and fix statistic ratio
This commit updates the model standings to ensure that they accurately reflect the performance of teams in the league. The changes include updating the calculations used to determine points and goals, as well as fixing an issue with the statistic ratio that was causing incorrect data to be displayed.

The following changes were made:

    Updated calculations used to determine points and goals
    Fixed issue with statistic ratio

These changes are designed to ensure that the model standings provide an accurate and reliable view of the league, and to help users make informed decisions when following the performance of their favorite teams.

Note: "fix" is used to indicate a bug fix or correction to existing functionality.